pyproject.toml
[project]
name = "pixiv-image-grouper"
version = "0.1.0"
description = "A tool to group Pixiv images"
requires-python = ">=3.12"
dependencies = [
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "pyinstaller",
    "ruff",
    "hatch",
    "pytest",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pixiv_image_grouper"]

[tool.hatch.build]
only-packages = true

[tool.ruff.format]
indent-style = "space"

[project.scripts]
pixiv-image-grouper = "pixiv_image_grouper.main:main"

[tool.hatch.envs.default.scripts]
build-exe = "pyinstaller src/pixiv_image_grouper/main.py --name pixiv-image-grouper --onefile"
test = "pytest tests"
typecheck = "mypy src/pixiv_image_grouper tests"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

[dependency-groups]
dev = [
    "mypy>=1.14.1",
]
